+++ cc3d/player5/ViewManager/SimpleViewManager.py
@@ -8,6 +8,7 @@ from cc3d.player5 import DefaultData
 import cc3d
 import datetime
 from cc3d.player5.Utilities.WebFetcher import WebFetcher
+from os import environ
 
 try:
     from cc3d.player5.Utilities.WebFetcherRequests import WebFetcherRequests
@@ -452,6 +453,11 @@ class SimpleViewManager(QObject):
         This function checks if new CC3D version is available
         :return:None
         """
+
+        
+        # checking if cc3d is running in nanohub. if it is do not check for updates (it'll be blocked by their firewall)
+        if 'NANOHUB_SIM' in environ:
+            return
 
         # here we decide whether the information about no new updates is displayed or not. For automatic update checks
         # this information should not be displayed. For manual update checks we need to inform the user